Don't wait by default when getting current activity.
Several places were incurring a 500ms delay when simply trying to get a
reference to the current activity. I fixed up the only place that I
could see that was relying on the sleep property of getCurrentActivity,
and changed the default so something that wouldn't slow down other code
unexpectedly.