Sleep at end of search loop, if view not found.
Many of the waits were sleeping at the start of the loop, before even
checking to see if the thing we were waiting for was already present.
By moving the waits to the end of each loop, we allow the loop to 
return more quickly if the element already exists.