Remove possibly unnecessary sleep.
This sleep happens before processing a list of views. It may be that 
we need to sleep to let the list of views settle? But the method that
uses this callback function does so in a loop, and it has its own sleep
at the end of that loop, so I think two calls to sleep is a bit too
much.